+++ tests/test_core.py
@@ -3620,12 +3620,13 @@ ok
     self.set_setting('MAIN_MODULE', 2)
     self.clear_setting('SIDE_MODULE')
     if auto_load:
-      # Normally we don't report undefined symbols when linking main modules but
-      # in this case we know all the side modules are specified on the command line.
-      # TODO(sbc): Make this the default one day
-      self.set_setting('ERROR_ON_UNDEFINED_SYMBOLS')
       self.emcc_args += main_emcc_args
       self.emcc_args.append('liblib.so')
+    else:
+      # When not auto loading (i.e. when not specifying the side module on the
+      # command line) we need to disable warnings on undefined symbols.
+      self.set_setting('WARN_ON_UNDEFINED_SYMBOLS', 0)
+
     if force_c:
       self.emcc_args.append('-nostdlib++')
 
@@ -4472,7 +4473,8 @@ res64 - external 64\n''', header='''
   @needs_dylink
   def test_dylink_dso_needed(self):
     def do_run(src, expected_output):
-      self.do_run(src + 'int main() { return test_main(); }', expected_output)
+      create_file('main.c', src + 'int main() { return test_main(); }')
+      self.do_runf('main.c', expected_output)
     self._test_dylink_dso_needed(do_run)
 
   @needs_dylink